+++ openeogeotrellis/collections/sentinel3.py
@@ -63,6 +63,7 @@ def main(product_type, native_resolution, bbox, from_date, to_date, band_names):
         layer_crs = layer.srdd.rdd().metadata().crs().epsgCode().get()
         layer.to_spatial_layer().save_stitched(f"/tmp/{product_type}_{from_date}_{to_date}.tif",
                                                crop_bounds=geopyspark.geotrellis.Extent(*bbox.reproject(layer_crs).as_wsen_tuple()))
+        return
 
         target_extent = Extent(*bbox.as_wsen_tuple())
         target_tileLayout = TileLayout(layoutCols=4, layoutRows=4, tileCols=256, tileRows=256)
@@ -265,7 +266,7 @@ def read_product(product, product_type, band_names, tile_size, limit_python_memo
             except Exception as e:
                 ex_type, ex_value, ex_traceback = sys.exc_info()
                 msg = f"Failed to read Sentinel-3 {product_type} {band_names} for {creo_path} and extent {layout_extent}. Error: {ex_value}"
-                logger.error(msg)
+                logger.error(msg, exc_info=True)
                 raise InternalException(msg)
 
             debug_mode = True
@@ -321,6 +322,14 @@ def create_s3_toa(product_type, creo_path, band_names, bbox_tile, digital_number
         geofile = 'geodetic_in.nc'
         lat_band = 'latitude_in'
         lon_band = 'longitude_in'
+    elif product_type in ["OL_2_LFR___", "OL_2_WFR___"]:
+        geofile = 'geo_coordinates.nc'
+        lat_band = 'latitude'
+        lon_band = 'longitude'
+    elif product_type == "SY_2_AOD___":
+        geofile = 'NTC_AOD.nc'
+        lat_band = 'latitude'
+        lon_band = 'longitude'
     else:
         raise ValueError(product_type)
 
@@ -451,6 +460,8 @@ def do_reproject(product_type, final_grid_resolution, creo_path, band_names,
                     return f"SDR_{band_name.split('_')[1]}"
             if product_type == SLSTR_PRODUCT_TYPE:
                 return band_name
+            if product_type in ["OL_2_LFR___", "OL_2_WFR___", "SY_2_AOD___"]:
+                return band_name
             raise ValueError(band_name)
 
         def readAndReproject(data_mask_,LUT):
@@ -551,19 +562,30 @@ def _read_latlonfile(bbox, latlon_file, lat_band="latitude", lon_band="longitude
                            ]
     to_drop = [x for x in potential_variables if x != lat_band and x != lon_band]  # saves memory
     # `open_dataarray` could allow for lazy loading, but is more complex and saves the same amount of memory
-    lat_lon_ds = xr.open_dataset(latlon_file, drop_variables=to_drop).astype("float32")
+    # decode_cf=True forces lat/lon bands from int32 to float64 by filling nans, scaling, and offsetting. This causes OOM errors.
+    # Instead, we do the scaling ourselves. Filling nans is not required for coordinates.
+    # Note that float32 causes us to lose precision, so we stick to int32.
+    lat_lon_ds = xr.open_dataset(latlon_file, drop_variables=to_drop, decode_cf=False)
+    lat_scale = lat_lon_ds[lat_band].attrs.get('scale_factor', 1.0)
+    lat_offset = lat_lon_ds[lat_band].attrs.get('add_offset', 0.0)
+    lon_scale = lat_lon_ds[lon_band].attrs.get('scale_factor', 1.0)
+    lon_offset = lat_lon_ds[lon_band].attrs.get('add_offset', 0.0)
 
-    xmin, ymin, xmax, ymax = bbox
+    # lat_lon_ds is scaled, so we need to inverse scale the bbox as well.
+    xmin_s, ymin_s, xmax_s, ymax_s = (bbox[0]/lon_scale)-lon_offset, (bbox[1]/lat_scale)-lat_offset, (bbox[2]/lon_scale)-lon_offset, (bbox[3]/lat_scale)-lat_offset
+    interpolation_margin_lat = (interpolation_margin/lat_scale)-lat_offset
+    interpolation_margin_lon = (interpolation_margin/lon_scale)-lon_offset
 
-    lat_mask = xr.apply_ufunc(lambda lat: (lat >= ymin - interpolation_margin) & (lat <= ymax + interpolation_margin), lat_lon_ds[lat_band])
-    lon_mask = xr.apply_ufunc(lambda lon: (lon >= xmin - interpolation_margin) & (lon <= xmax + interpolation_margin), lat_lon_ds[lon_band])
+    lat_mask = xr.apply_ufunc(lambda lat: (lat >= ymin_s - interpolation_margin_lat) & (lat <= ymax_s + interpolation_margin_lat), lat_lon_ds[lat_band])
+    lon_mask = xr.apply_ufunc(lambda lon: (lon >= xmin_s - interpolation_margin_lon) & (lon <= xmax_s + interpolation_margin_lon), lat_lon_ds[lon_band])
     data_mask = lat_mask & lon_mask
 
 
     # Create the coordinate arrays for latitude and longitude
     ## Coordinated referring to the CENTER of the pixel
-    lat_orig = lat_lon_ds[lat_band].where(data_mask,drop=True).values
-    lon_orig = lat_lon_ds[lon_band].where(data_mask,drop=True).values
+    # We can rescale to float64 from this point since we are working with smaller arrays.
+    lat_orig = lat_lon_ds[lat_band].where(data_mask, drop=True).values*lat_scale+lat_offset
+    lon_orig = lat_lon_ds[lon_band].where(data_mask, drop=True).values*lon_scale+lon_offset
     lat_lon_ds.close()
 
     if lat_orig.size == 0 or lon_orig.size == 0:
@@ -576,8 +598,8 @@ def _read_latlonfile(bbox, latlon_file, lat_band="latitude", lon_band="longitude
         passing_date_line = True
         # self.lon_orig=np.where(self.lon_orig<0, self.lon_orig+360, self.lon_orig) #change grid from -180->180 to 0->360
 
-    x_min, x_max = np.min(lon_orig), np.max(lon_orig)
-    y_min, y_max = np.min(lat_orig), np.max(lat_orig)
+    x_min, x_max = np.nanmin(lon_orig), np.nanmax(lon_orig)
+    y_min, y_max = np.nanmin(lat_orig), np.nanmax(lat_orig)
     bbox_original = [x_min, y_min, x_max, y_max]
     lon_flat = lon_orig.flatten()
     lat_flat = lat_orig.flatten()
@@ -732,14 +754,14 @@ if __name__ == '__main__':
 
     lat_lon_bbox = [2.535352308127358, 50.57415247573394, 5.713651867060349, 51.718230797191836]
     lat_lon_bbox = [9.944991786580573, 45.99238819027832, 12.146700668591137, 47.27025711819684]
-    lat_lon_bbox = [-128.4272367635350349, 49.7476186207236424, -126.9726189291401113, 50.8176823149911527]
+    lat_lon_bbox = [129.4502384682777, -18.161201081701407, 138.1236995771855, -10.287173762653026]
     # lat_lon_bbox = [0.0, 50.0, 5.0, 55.0]
-    from_date = "2018-03-12T00:00:00Z"
-    to_date = from_date
-    band_names = ["LST_in:LST"]
+    from_date = "2024-01-01T00:00:00Z"
+    to_date = "2024-01-01T02:00:00Z"
+    band_names = ["NTC_AOD:Surface_reflectance_440"]
 
-    product_type = SLSTR_PRODUCT_TYPE
-    native_resolution = 0.008928571428571
-    bbox = BoundingBox.from_wsen_tuple(lat_lon_bbox, crs=4326).reproject(32609)
+    product_type = "SY_2_AOD___"
+    native_resolution = 0.040178571
+    bbox = BoundingBox.from_wsen_tuple(lat_lon_bbox, crs=4326).reproject(32753)
 
     main(product_type, native_resolution, bbox, from_date, to_date, band_names)