+++ tests/integrations/test_calrissian.py
@@ -1,3 +1,4 @@
+from pathlib import Path
 from typing import Dict
 from unittest import mock
 
@@ -268,6 +269,36 @@ class TestCalrissianS3Result:
         bucket, key = s3_output
         result = CalrissianS3Result(s3_bucket=bucket, s3_key=key)
         assert result.read(encoding="utf-8") == "Howdy, Earth!"
+
+    def test_generate_presigned_url(self, s3_output, monkeypatch):
+        monkeypatch.setenv("SWIFT_URL", "https://s3.example.com")
+        bucket, key = s3_output
+        result = CalrissianS3Result(s3_bucket=bucket, s3_key=key)
+        assert result.generate_presigned_url() == dirty_equals.IsStr(
+            regex=r"https://s3.example.com/the-bucket/path/to/output.txt\?AWSAccessKeyId=.*"
+        )
+
+    def test_generate_public_url(self, s3_output, monkeypatch):
+        monkeypatch.setenv("SWIFT_URL", "https://s3.example.com")
+        bucket, key = s3_output
+        result = CalrissianS3Result(s3_bucket=bucket, s3_key=key)
+        assert result.generate_public_url() == "https://s3.example.com/the-bucket/path/to/output.txt"
+
+    def test_download(self, s3_output, tmp_path):
+        bucket, key = s3_output
+        result = CalrissianS3Result(s3_bucket=bucket, s3_key=key)
+        path = tmp_path / "result.data"
+        result.download(path)
+        assert path.read_bytes() == b"Howdy, Earth!"
+
+    def test_download_to_dir(self, s3_output, tmp_path):
+        bucket, key = s3_output
+        result = CalrissianS3Result(s3_bucket=bucket, s3_key=key)
+        folder = tmp_path / "data"
+        folder.mkdir(parents=True)
+        path = result.download(target=folder)
+        assert path.relative_to(folder) == Path("output.txt")
+        assert path.read_bytes() == b"Howdy, Earth!"
 
 
 class TestCwlSource: